Implemented some of the internals, untested
Some of the internals for the retrieval function and a 'foreach'
function have been implemented. They are completely untested as
currently there is no database to read from (and some other basic
functionality is missing).

The internals are modelled after the python implementation of the CDB
interface, available in the 'docs/' folder. It is really quite readable,
much better than the specifications for format.

Also added:

* Added unit test functions/framework, which is quite trivial at the
moment.
* Added a spartan manual page, spartan as there is not much
functionality in the cdb program to document.
* A 'getopt' function has been added to 'main.c'.
* A library for CDB is built as part of the build process
* The documentation has been improved slightly, and more project goals
have been added.