Implement dump and read functions
The major change in this commit is that there is now some basic
functionality implemented, the dump database and dump keys functions
have been added, as well as the read key from database function. Both
functions were tested by generating a database with the python script
and using that to validate that it can retrieve keys correctly.

* A wish list of items and possible improvements has been added
* More command line options have been added
* Some To Do items have been added and some ticked off, the major new
item is multiple key retrieval needs to be dealt with, it is perfectly
valid for a database to have multiple entries with the same key, so a
way of extracting that information needs to be added, a minor
complication.