Added creation and statistics collection (working)
This is the first version of the program that can be considered working.
There is still more to do, but it is mostly polishing and not adding
functionality to the core library. The main changes in this commit are
that database creation now works, and some statistics for a database can
be printed. There are also some minor cosmetic changes that account for
many diffs (such as changing references to 'uint32_t' to a typedef of
'word_t' within the library).

Other changes:

* On Windows the file streams should be (re)opened in binary mode
instead of text mode.
* Duplicate keys can be retrieved (untested)
* The documentation has been updated to reflect changes, the format is
also mostly documented in the 'readme.md' file. Some ASCII text diagrams
would help, as would rewriting some of the prose. It is more awkward
than it needs to be in an attempt to be accurate.

Main To Do Items:

* The query format needs improving, arbitrary keys cannot be retrieved
at the moment (only keys up to a size of 256 bytes, excluding an ASCII
NUL can be retrieved, the same goes for items to be displayed via the
query, but not via the database dump).
* Collect more statistics.
* Address some of the allocation concerns within the 'cdb.c' file.
* Experiment with the library!