get bug fix, unit tests, i/o, & more!
* Unit tests have been added as a built in self test in the C library,
which can be compiled out by defining NDEBUG. The utility of unit test
is proved again as they found a bug in the library. The unit tests use
the same callback mechanism as the rest of the library, so do not require
functionality not available on a non-hosted platform to be present for
compilation. More things need to be tested, mostly to do with the
wrapper program and duplicate key retrieval/storage.
* A bug has been fixed in the 'cdb_get_record' function where the
comparison used only worked for the first 256 bytes
* The input parsing in the test application.
* The first hash table, which is of a fixed length and quite small, can
be held in memory to speed things up. This is selected by a compile time
option which is by default disabled.
* A query mode has been added to the main application, allowing quick
querying of NUL terminated ASCII strings (the rest of the library can
fully deal with binary data, this is the exception).
* The manual page has been improved, but it is still small, with most of
the documentation being placed in 'readme.md'. The text output of the
help option has been cut back to keep the C program nice and small, it
was getting too large.
* A test script, called 't', has been added, it requires '/bin/sh' to be
installed. It does some things that the Built In Self Test (BIST) does not.
The BIST needs improvement as well, but this script is easier to modify.