Removed docs & cdb.1, more t, more checks, docs
* The 'cdb.1' man page has been removed and is now generated from the
'readme.md' file, which is in the process of being rewritten to look
more like a manual page.
* The 'docs/' folder has been removed, it is no longer needed.
References to the webpages from which the documents/code was retrieved
still remain the project.
* More tests have been added to the 't' program.
* There are now two new makefile targets; 'install' and 'dist'
* More checks on the data that is being read in have been added, the
database driver no longer blindly trusts the database but range checks
and other sanity checks. If the checks fail, the database connection
object is invalidated, and subsequent function calls fail to run.
* The TODO list has been streamlined.