API changes, calculate more stats, error # ...
This is a fairly big commit!

The API has been changed with references to 'cdb_word_t', I think it has
been rationalized and made tidier. The functions 'cdb_read' and
'cdb_seek' have been added to the API, whilst the function
'cdb_get_file' has been removed.

Internally to the library error numbers are used so what caused a
problem can be determined to a limited degree instead of just 'an
error'.

More statistics are printed out, and more things are checked for in the
file format to make things add up.

More tests have been added to the built in test bench, duplicate keys
are now tested for. The test script 't' is now more structured as well,
and tests for slightly more things, as well as some expected failure
conditions.

Other changes:
* The documentation, 'readme.md' has been updated. It is making for
quite a comprehensive manual and project page.
* New build targets and behavior; a shared library is built and part of
the dist/install targets, as well as the 'cdb.h' header file, which was
missing.
* A version number can be set via the build system. It may need
changing/updating.

Possible changes:
* There are only a few TODO comments left to do.
* After the TODO comments are done, it will mostly be a case of stress
testing, performance testing, and making sure that 16 and 64 bit
versions of the library - if the user so chooses to compile the library
as so, work.