Added 'cdb_get_count', improved test framework
* A database sanity check has been moved from 'main.c' to inside the
'cdb.c', it is called each time the database is opened.
* 'cdb_get_count' has been added, which allows the number of items in a
database with the same key to be queried. The documentation has been
updated to reflect this.
* The test framework has been improved, and simplified at the same time.
The automatic generation of test cases can now take into account
possible duplicates that are generated. The test framework also works
for the 16-bit version of the library. The 64-bit version does not work
yet. Whilst most operations appear to work when run manually for the
64-bit version, the built in tests do not complete correctly.
* There have been some minor style changes.