Fixed 64 bit version, broke 16
The 64-bit version of the library has been fixed, the problem was some
implicit integer conversion being done on a byte value. The 16-bit
version of the library no longer seems to pass its internal tests, which
is a shame, it was before. The build system has been improved, various
things can be disabled via the make file, and the 16, 32 and 64 bit
versions can be build via the build script. Other than that, some TODO
comments have been resolved/added, a bug in record retrieval for the
'-r' option fixed, the way version numbers are dealt with has changed (a
function for retrieving the library version has been added), and there
some other minor style changes.