Added '-v' and '-T', fixed file open bug
* A bug has been removed whereby if a file failed to open cleanup code
would trigger an assert because an error was not properly signalled.
* Options '-v' to increase verbosity have been added (along with some
new messages) and '-T' to specify the name of a temporary file to use.
This can be used during file creation to replace a file that is
currently in use atomically. The database is recreated into the
temporary file and then it is atomically replaced with 'rename'.
'tmpnam' can be used internally by default if no file is used with '-T',
however that may cause security problems so it is disabled by default,
instead it just (non-atomically) updated the file in place which may
cause consistency problems if no '-T' is specified and 'USE_TMPNAM' is
defined as '0'.
* The help message has examples and options put into it, again.