A brand new API!
Well, a version that is incompatible in minor ways with 1.X.X version,
so the MAJOR version is being bumped to indicate this. The changes in
this commit are:

* Removing white space from the end of all the files
* Making the arguments to a few functions 'const' when they can be
(breaking change).
* Removing the 'cdb_read_word' function, you either want to perform a
'cdb_read' or a 'cdb_read_word_pair'.
* 'cdb_version' has been renamed to 'cdb_get_version' and returns a
'uint32_t' not a 'unsigned long'.
* A few new ideas have been added to the 'readme.md' file for
integrating this library into other project (my 'pickle' and 'eweb'
projects).
* The makefile has been cleaned up a little.