Improved performance: API changes!
The API changes continue, which will go into the new version. Not only
that but the command line interface has changed somewhat as well, making
it completely incompatible with version 2.0 and prior of the program.

* Some more checks are done on API functions to make sure they are not
used incorrectly.
* The API is now more consistent, everything returns an integer to
indicate failure.
* Fewer seeks are performed and fewer allocations (although it will
allocate more than it needs to). The file seek operation really slow
things down, reducing the number of them means the file buffer does not
have to be discarded.