Can now set file format size by command line option
Multiple CDB file formats (16, 32 [default] and 64-bit) can be created
with the same library without recompiling it, it can now be specified at
run time instead of compile time. Some other things need to be changed to
accommodate this, some of the tests and the documentation in 'readme.md'
need to be updated to reflect this.