failing test: offset
In this commit:
- An offset for the file start position can be added so the CDB file can
be embedded within other files. An option has been added so that the
offset into a file can be specified by the command line.
- A function called 'cdb_next' has been added, but not tested, for
accessing values with the same key quicker.
- Some TODO comments have been removed, some added.

BUG:

* For the 64-bit version specifying the offset does not seem to
work/causes problems. This needs fixing. The tests for the new offset
mechanism fail as a consequence of this, the 16-bit and 32-bit versions
pass.